send-feedback: more logs property and safeHideLoader property
This commit add two new property's for logs:
logsFetched will be set true if logs were fetched by this element,
and if this is set that means logFiles property should have all the logs
read that you pushed to logs property's array.

other than that a new function is added called safeHideLoader that will
only hide loader if it was shown.