send-feedback: Fix a bug where show-cancel-button attribute wasn't reflected.
This is a pretty weird bug, and it occurs because we rely on TemplateManager
to sync everything from the attribute to the state of the element. For example,
if the title attribute was set, we have {{ title }} in the template so it gets
reflected but we don't have such code for show-cancel-button. Hence, we need to
handle this in connectedCallback.